+++ c/mpi/collective/osu_reduce.c
@@ -16,7 +16,7 @@ int main(int argc, char *argv[])
     double latency = 0.0, t_start = 0.0, t_stop = 0.0;
     double timer=0.0;
     double avg_time = 0.0, max_time = 0.0, min_time = 0.0;
-    float *sendbuf, *recvbuf;
+    int *sendbuf, *recvbuf;
     int po_ret;
     int errors = 0, local_errors = 0;
     size_t bufsize;
@@ -70,12 +70,12 @@ int main(int argc, char *argv[])
         exit(EXIT_FAILURE);
     }
     check_mem_limit(numprocs);
-    options.min_message_size /= sizeof(float);
+    options.min_message_size /= sizeof(int);
     if (options.min_message_size < MIN_MESSAGE_SIZE) {
         options.min_message_size = MIN_MESSAGE_SIZE;
     }
 
-    bufsize = sizeof(float) * (options.max_message_size / sizeof(float));
+    bufsize = sizeof(int) * (options.max_message_size / sizeof(int));
     if (allocate_memory_coll((void**)&recvbuf, bufsize,
                 options.accel)) {
         fprintf(stderr, "Could Not Allocate Memory [rank %d]\n", rank);
@@ -83,7 +83,7 @@ int main(int argc, char *argv[])
     }
     set_buffer(recvbuf, options.accel, 1, bufsize);
 
-    bufsize = sizeof(float) * (options.max_message_size / sizeof(float));
+    bufsize = sizeof(int) * (options.max_message_size / sizeof(int));
     if (allocate_memory_coll((void**)&sendbuf, bufsize,
                 options.accel)) {
         fprintf(stderr, "Could Not Allocate Memory [rank %d]\n", rank);
@@ -94,7 +94,7 @@ int main(int argc, char *argv[])
     print_preamble(rank);
     omb_papi_init(&papi_eventset);
 
-    for (size = options.min_message_size; size * sizeof(float) <=
+    for (size = options.min_message_size; size * sizeof(int) <=
             options.max_message_size; size *= 2) {
         if (size > LARGE_MESSAGE_SIZE) {
             options.skip = options.skip_large;
@@ -102,7 +102,7 @@ int main(int argc, char *argv[])
         }
 
         omb_graph_allocate_and_get_data_buffer(&omb_graph_data,
-                &omb_graph_options, size * sizeof(float), options.iterations);
+                &omb_graph_options, size * sizeof(int), options.iterations);
         MPI_CHECK(MPI_Barrier(MPI_COMM_WORLD));
 
         timer=0.0;
@@ -121,7 +121,7 @@ int main(int argc, char *argv[])
                 }
                 for (j = 0; j < options.warmup_validation; j++) {
                     MPI_CHECK(MPI_Barrier(MPI_COMM_WORLD));
-                    MPI_CHECK(MPI_Reduce(sendbuf, recvbuf, size, MPI_FLOAT,
+                    MPI_CHECK(MPI_Reduce(sendbuf, recvbuf, size, MPI_INT,
                                 MPI_SUM, 0, MPI_COMM_WORLD ));
                 }
             }
@@ -129,7 +129,7 @@ int main(int argc, char *argv[])
 
             t_start = MPI_Wtime();
 
-            MPI_CHECK(MPI_Reduce(sendbuf, recvbuf, size, MPI_FLOAT, MPI_SUM, 0,
+            MPI_CHECK(MPI_Reduce(sendbuf, recvbuf, size, MPI_INT, MPI_SUM, 0,
                         MPI_COMM_WORLD ));
             t_stop=MPI_Wtime();
 
@@ -148,7 +148,7 @@ int main(int argc, char *argv[])
                 }
             }
         }
-        omb_papi_stop_and_print(&papi_eventset, size * sizeof(float));
+        omb_papi_stop_and_print(&papi_eventset, size * sizeof(int));
         latency = (double)(timer * 1e6) / options.iterations;
 
         MPI_CHECK(MPI_Reduce(&latency, &min_time, 1, MPI_DOUBLE, MPI_MIN, 0,
@@ -164,10 +164,10 @@ int main(int argc, char *argv[])
         }
 
         if (options.validate) {
-            print_stats_validate(rank, size * sizeof(float), avg_time, min_time,
+            print_stats_validate(rank, size * sizeof(int), avg_time, min_time,
                     max_time, errors);
         } else {
-            print_stats(rank, size * sizeof(float), avg_time, min_time,
+            print_stats(rank, size * sizeof(int), avg_time, min_time,
                     max_time);
         }
         if (options.graph && 0 == rank) {
